requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy==2.2.1
pydantic==2.10.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
